-Changed int to 3 byte -Removed extra input delay -Cleaned up API thread safety -Added Bag slots free info

+++ WowAutomater/BuffSpell.cs
@@ -13,10 +13,10 @@ namespace ClassicWowNeuralParasite
 
         public BuffSpell (VirtualKeyCode hotKey,
                           BuffType buff,
-                          ushort manaCost = 0,
+                          uint manaCost = 0,
                           double cooldownTime = 0,
                           double healthPercentage = 100,
-                          ushort level = 1,
+                          uint level = 1,
                           bool useOnce = false) : base(hotKey, manaCost, cooldownTime, healthPercentage, level, useOnce)
         {
             Buff = buff;
